replace hardcoded currency on header cart button
